APIを飲み会と支払いに関して分けました
