+++ src/main/java/jp/co/biglobe/warikan/api/payment/PartyRegistrationApi.java
@@ -1,0 +1,28 @@
+package jp.co.biglobe.warikan.api.payment;
+
+import jp.co.biglobe.warikan.datasource.PartyRepository;
+import jp.co.biglobe.warikan.domain.repository.IPartyRepository;
+import jp.co.biglobe.warikan.service.PartyApplicationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class PartyRegistrationApi {
+    private IPartyRepository partyRepository = new PartyRepository();
+    private PartyApplicationService partyApplicationService = new PartyApplicationService(partyRepository);
+
+    @GetMapping("/party/register")
+    public PartyRegistrationResponse registerParty(PartyRegistrationRequest request) {
+
+        double largeRatio = request.getLargeRatio();
+        double mediumRatio = request.getMediumRatio();
+        double smallRatio = request.getSmallRatio();
+
+        int createdPartyId = partyApplicationService.registerParty(largeRatio, mediumRatio, smallRatio);
+
+        return new PartyRegistrationResponse(createdPartyId);
+
+    }
+}
